added data structures and removed useless code

+++ fuzzy_compare.h
@@ -5,7 +5,7 @@
 namespace util_lib{
 	unsigned int fuzzyCompare(const std::string & w1, const std::string & w2);
 	unsigned int fuzzyCompare2(const std::string & w1, const std::string & w2);
-	unsigned int fuzzyCompare3(const std::string & w1, const std::string & w2,int * startIncluding ,int *endIncluding );
+	unsigned int fuzzyCompare3(const std::string & w1, const std::string & w2,long * startIncluding ,long *endIncluding );
 }
 
 #endif
